+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Save license when a video is created
 - Show an error when a resource has been deleted
+- Allow depositedfiles to have unicode characters in their filename
 
 ### Changed